+++ include/cli/split.h
@@ -39,72 +39,196 @@ namespace cli
 namespace detail
 {
 
-
-inline void split(std::vector<std::string>& strs, const std::string& input)
+class Text
 {
-    enum class State { space, word, sentence };
-    State state = State::space;
+public:
+    explicit Text(const std::string& _input) : input(_input)
+    {
+    }
+    void SplitInto(std::vector<std::string>& strs)
+    {
+        Reset();
+        for (char c: input)
+            Eval(c);
+        RemoveEmptyEntries();
+        splitResult.swap(strs); // puts the result back in strs
+    }
+private:
+    void Reset()
+    {
+        state = State::space;
+        prev_state = State::space;
+        sentence_type = SentenceType::double_quote;
+        splitResult.clear();
+    }
 
-    strs.clear();
-
-    for (char c: input)
+    void Eval(char c)
     {
         switch(state)
         {
             case State::space:
-                if (c == ' ' || c == '\t' || c == '\n')
-                {
-                    // do nothing
-                }
-                else if (c == '"')
-                {
-                    state = State::sentence;
-                    strs.push_back("");
-                }
-                else
-                {
-                    state = State::word;
-                    strs.push_back(std::string(1, c));
-                }
+                EvalSpace(c);
                 break;
             case State::word:
-                if (c == ' ' || c == '\t' || c == '\n')
-                {
-                    state = State::space;
-                }
-                else if (c == '"')
-                {
-                    state = State::sentence;
-                    strs.push_back("");
-                }
-                else
-                {
-                    assert(!strs.empty());
-                    strs.back() += c;
-                }
+                EvalWord(c);
                 break;
             case State::sentence:
-                if (c == '"')
-                {
-                    state = State::space;
-                }
-                else
-                {
-                    assert(!strs.empty());
-                    strs.back() += c;
-                }
+                EvalSentence(c);
+                break;
+            case State::escape:
+                EvalEscape(c);
+                break;
         }
     }
 
-    // remove null entries from the vector:
-    strs.erase(
-        std::remove_if(
-            strs.begin(),
-            strs.end(),
-            [](const std::string& s){ return s.empty(); }
-        ),
-        strs.end()
-    );
+    void EvalSpace(char c)
+    {
+        if (c == ' ' || c == '\t' || c == '\n')
+        {
+            // do nothing
+        }
+        else if (c == '"' || c == '\'')
+        {
+            NewSentence(c);
+        }
+        else if (c == '\\')
+        {
+            // This is the case where the first character of a word is escaped.
+            // Should come back into the word state after this.
+            prev_state = State::word;
+            state = State::escape;
+            splitResult.push_back("");
+        }
+        else
+        {
+            state = State::word;
+            splitResult.push_back(std::string(1, c));
+        }
+    }
+
+    void EvalWord(char c)
+    {
+        if (c == ' ' || c == '\t' || c == '\n')
+        {
+            state = State::space;
+        }
+        else if (c == '"' || c == '\'')
+        {
+            NewSentence(c);
+        }
+        else if (c == '\\')
+        {
+            prev_state = state;
+            state = State::escape;
+        }
+        else
+        {
+            assert(!splitResult.empty());
+            splitResult.back() += c;
+        }
+    }
+
+    void EvalSentence(char c)
+    {
+        if (c == '"' || c == '\'')
+        {
+            auto new_type = c == '"' ? SentenceType::double_quote : SentenceType::quote;
+            if (new_type == sentence_type)
+                state = State::space;
+            else
+            {
+                assert(!splitResult.empty());
+                splitResult.back() += c;
+            }
+        }
+        else if (c == '\\')
+        {
+            prev_state = state;
+            state = State::escape;
+        }
+        else
+        {
+            assert(!splitResult.empty());
+            splitResult.back() += c;
+        }
+    }
+
+    void EvalEscape(char c)
+    {
+        assert(!splitResult.empty());
+        if (c != '"' && c != '\'' && c != '\\')
+            splitResult.back() += "\\";
+        splitResult.back() += c;
+        state = prev_state;
+    }
+
+    void NewSentence(char c)
+    {
+        state = State::sentence;
+        sentence_type = ( c == '"' ? SentenceType::double_quote : SentenceType::quote);
+        splitResult.push_back("");
+    }
+
+    void RemoveEmptyEntries()
+    {
+        // remove null entries from the vector:
+        splitResult.erase(
+            std::remove_if(
+                splitResult.begin(),
+                splitResult.end(),
+                [](const std::string& s){ return s.empty(); }
+            ),
+            splitResult.end()
+        );
+    }
+
+    enum class State { space, word, sentence, escape };
+    enum class SentenceType { quote, double_quote };
+    State state = State::space;
+    State prev_state = State::space;
+    SentenceType sentence_type = SentenceType::double_quote;
+    const std::string input;
+    std::vector<std::string> splitResult;
+};
+
+// Split the string input into a vector of strings.
+// The original string is split where there are spaces.
+// Quotes and double quotes can be used to indicate a substring that should not be splitted
+// (even if it contains spaces)
+
+//          split(strs, "");                => empty vector
+//          split(strs, " ");               => empty vector
+//          split(strs, "  ");              => empty vector
+//          split(strs, "\t");              => empty vector
+//          split(strs, "  \t \t     ");    => empty vector
+
+//          split(strs, "1234567890");      => <"1234567890">
+//          split(strs, "  foo ");          => <"foo">
+//          split(strs, "  foo \t \t bar \t");  => <"foo","bar">
+
+//          split(strs, "\"\"");            => empty vector
+//          split(strs, "\"foo bar\"");     => <"foo bar">
+//          split(strs, "    \t\t \"foo \tbar\"     \t");   => <"foo \tbar">
+//          split(strs, " first   \t\t \"foo \tbar\"     \t last"); => <"first","foo \tbar","last">
+//          split(strs, "first\"foo \tbar\"");  => <"first","foo \tbar">
+//          split(strs, "first \"'second' 'thirdh'\""); => <"first","'second' 'thirdh'">
+
+//          split(strs, "''");      => empty vector
+//          split(strs, "'foo bar'");   => <"foo bar">
+//          split(strs, "    \t\t 'foo \tbar'     \t"); => <"foo \tbar">
+//          split(strs, " first   \t\t 'foo \tbar'     \t last"); => <"first","foo \tbar","last">
+//          split(strs, "first'foo \tbar'"); => <"first","foo \tbar">
+//          split(strs, "first '\"second\" \"thirdh\"'"); => <"first","\"second\" \"thirdh\"">
+
+//          split(strs, R"("foo\"bar")"); // "foo\"bar" => <"foo"bar">
+//          split(strs, R"('foo\'bar')"); // 'foo\'bar' => <"foo'bar">
+//          split(strs, R"("foo\bar")"); // "foo\bar" => <"foo\bar">
+//          split(strs, R"("foo\\"bar")"); // "foo\\"bar" => <"foo\"bar">
+
+inline void split(std::vector<std::string>& strs, const std::string& input)
+{
+    Text sentence(input);
+    sentence.SplitInto(strs);
 }
 
 } // namespace detail