+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Current Version: 1.1.1-dirty
  - Handle CTRL-D (EOF) on linux
  - Handle CTRL-D, CTRL-Z and CTRL-C on windows
  - Fix Backspace from remote terminal (issue #52)
+ - Escaping and sentence support with quote and double quote
 
 ## [1.1.1] - 2019-09-16